fix: budget/ex type err, by zero err

+++ src/main/java/com/travel/rate/dto/travel/ReqTravelDTO.java
@@ -7,6 +7,6 @@ import lombok.Data;
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 @Data
 public class ReqTravelDTO {
-    Long budget;
+    int budget;
     Long continent_id;
 }
